se guarda el producto
