+++ assets/js/productos.js
@@ -1,90 +1,146 @@
-$(document).ready(function() {
-  $('#table_productos').DataTable({
-      "ajax": {
-          "url": "controllers/productosController.php?option=listar",
-          "dataSrc": "",
-          "error": function(xhr, error, code){
-              console.log(xhr.responseText);
-          }
-      },
-      "columns": [
-          {"data": "id_producto"},
-          {"data": "codigo_producto"},
-          {"data": "descripcion"},
-          {"data": "precio_compra"},
-          {"data": "precio_venta"},
-          {
-            "data": "imagen",
-            "render": function(data, type, row) {
-                if (data) {
-                    // imagen para que sea relativa a la raíz del servidor web
-                    var relativePath = data.replace("C:\\xampp\\htdocs\\venta\\", "").replace(/\\/g, "/");
-                    return '<img src="' + relativePath + '" width="50" height="50"/>';
+document.addEventListener('DOMContentLoaded', function () {
+    // Elementos del DOM
+    const formProductos = document.getElementById('frmProductos');
+    const btnRecibido = document.getElementById('btn-Recibido');
+    const selectSede = document.getElementById('selectSede');
+    const modalMensajeRol = document.getElementById('modalMensajeRol');
+    const guardarModal = document.getElementById('guardarModal');
+    const cerrarModal = document.getElementById('cerrarModal');
+    const table_productos = document.querySelector('#table_productos tbody');
+
+    // Función para cargar los productos desde el backend
+    function cargarProductos() {
+        axios.get(`controllers/productosController.php?option=listarProductos&id_caja=${idSede}&rolUsuario=${rolUsuario}`)
+            .then(response => {
+                const data = response.data;
+                if (data.productos) {
+                    mostrarProductos(data.productos);
                 } else {
-                    return '<img src="uploads/default.png" width="50" height="50"/>'; // Ruta a una imagen por defecto
+                    console.error('Error al obtener los productos:', data.error);
                 }
-            }
-        },
-        
-        
-          {"data": "existencia"},
-      ]
-  });
+            })
+            .catch(error => console.error('Error al cargar productos:', error));
+    }
 
-  
-});
+    // Mostrar productos en la tabla
+    function mostrarProductos(productos) {
+        table_productos.innerHTML = '';
+        productos.forEach(producto => {
+            const imagenProducto = producto.Imagen ? producto.Imagen : 'uploads/default.png';
 
+            const row = document.createElement('tr');
+            row.innerHTML = `
+                <td>${producto.Id}</td>
+                <td>${producto.Barcode}</td>
+                <td>${producto.Descripción}</td>
+                <td>${producto.Precio_Compra}</td>
+                <td>${producto.Precio_Venta}</td>
+                <td>${producto.Proveedor}</td>
+                <td>
+                    <img src="${imagenProducto}" alt="Imagen del Producto" width="50" 
+                    onerror="this.src='uploads/default.png';" />
+                </td>
+                <td>${producto.Cantidad}</td>
+            `;
+            table_productos.appendChild(row);
+        });
+    }
 
-function deleteProducto(id) {
-  $.ajax({
-      url: 'controllers/productosController.php?option=delete&id=' + id,
-      type: 'GET',
-      success: function(response) {
-          var res = JSON.parse(response);
-          if (res.tipo === 'success') {
-              $('#table_productos').DataTable().ajax.reload();
-              alert(res.mensaje);
-          } else {
-              alert(res.mensaje);
-          }
-      }
-  });
-}
+    // Función para cargar los proveedores en el select
+    function cargarProveedores() {
+        axios.get('controllers/productosController.php?option=listarProveedores')
+            .then(response => {
+                const data = response.data;
+                const selectProveedor = document.getElementById('id_empresa');
 
-function editProducto(id) {
-  $.ajax({
-      url: 'controllers/productosController.php?option=edit&id=' + id,
-      type: 'GET',
-      success: function(response) {
-          var data = JSON.parse(response);
-          $('#id_product').val(data.codproducto);
-          $('#barcode').val(data.codigo);
-          $('#nombre').val(data.descripcion);
-          $('#precio').val(data.precio_compra); 
-          $('#stock').val(data.existencia);
-          $('#modalProducto').modal('show');
-      }
-  });
-}
+                if (data.proveedores) {
+                    selectProveedor.innerHTML = '<option value="">Seleccione un proveedor</option>';
+                    data.proveedores.forEach(proveedor => {
+                        selectProveedor.innerHTML += `<option value="${proveedor.id_empresa}">${proveedor.razon_social}</option>`;
+                    });
+                } else {
+                    console.error('Error al obtener proveedores:', data.error);
+                }
+            })
+            .catch(error => console.error('Error al cargar proveedores:', error));
+    }
 
-$('#formProducto').submit(function(e) {
-  e.preventDefault();
-  var formData = new FormData(this);
-  $.ajax({
-      url: 'controllers/productosController.php?option=save',
-      type: 'POST',
-      data: formData,
-      contentType: false,
-      processData: false,
-      success: function(response) {
-          var res = JSON.parse(response);
-          if (res.tipo === 'success') {
-              $('#table_productos').DataTable().ajax.reload();
-              $('#modalProducto').modal('hide');
-              alert(res.mensaje);
-          } else {
-              alert(res.mensaje);
-          }
-      }
-  });
+    // Manejar el evento de envío del formulario principal
+    formProductos.addEventListener('submit', function (e) {
+        e.preventDefault(); // Evitar que el formulario recargue la página
+
+        const formData = new FormData(formProductos);
+
+        // Enviar los datos al backend para registrar el producto
+        axios.post('controllers/productosController.php?option=registrarProducto', formData)
+            .then(response => {
+                const data = response.data;
+
+                if (data.tipo === 'success') {
+                    Swal.fire('Éxito', data.mensaje, 'success');
+                    cargarProductos(); // Recargar los productos para mostrar el nuevo
+                } else {
+                    Swal.fire('Error', data.mensaje, 'error');
+                }
+            })
+            .catch(error => {
+                console.error('Error al registrar el producto:', error);
+                Swal.fire('Error', 'Hubo un problema al guardar el producto.', 'error');
+            });
+    });
+
+    // Abrir el modal al presionar "Recibido"
+    btnRecibido.addEventListener('click', function () {
+        modalMensajeRol.style.display = 'flex'; // Muestra el modal
+    });
+
+    // Manejar el evento "Guardar" en el modal
+    guardarModal.addEventListener('click', function () {
+        const idCaja = selectSede.value; // Valor seleccionado en el modal
+        if (!idCaja) {
+            alert('Por favor seleccione una sede antes de continuar.');
+            return;
+        }
+
+        // Crear FormData a partir del formulario
+        const formData = new FormData(formProductos);
+
+        // Agregar el valor de `id_caja` del modal al FormData
+        formData.append('id_caja', idCaja);
+
+        // Verificar los datos enviados (opcional para depuración)
+        for (let [key, value] of formData.entries()) {
+            console.log(`${key}: ${value}`);
+        }
+
+        // Enviar datos al backend
+        axios.post('controllers/productosController.php?option=registrarProducto', formData)
+            .then(response => {
+                const data = response.data;
+
+                if (data.tipo === 'success') {
+                    Swal.fire('Éxito', data.mensaje, 'success');
+                    modalMensajeRol.style.display = 'none'; // Oculta el modal
+                    formProductos.reset(); // Limpia el formulario
+                    cargarProductos(); // Recarga la tabla de productos
+                } else {
+                    Swal.fire('Error', data.mensaje, 'error');
+                }
+            })
+            .catch(error => {
+                console.error('Error al guardar el producto:', error);
+                Swal.fire('Error', 'Hubo un problema al guardar el producto.', 'error');
+            });
+    });
+
+    // Cerrar el modal
+    cerrarModal.addEventListener('click', function () {
+        modalMensajeRol.style.display = 'none';
+    });
+
+    // Cargar productos y proveedores al cargar la página
+    if (table_productos) {
+        cargarProductos();
+    }
+    cargarProveedores();
 });
