+++ controllers/comprasController.php
@@ -1,188 +1,71 @@
 <?php
-$option = isset($_GET['option']) ? $_GET['option'] : '';
-require_once '../models/compras.php';
-$compras = new Compras();
-$id_user = $_SESSION['idusuario'];
-$option = isset($_GET['option']) ? $_GET['option'] : '';
-error_log("Opción recibida: $option");
-
-
+$option = (empty($_GET['option'])) ? '' : $_GET['option'];
+require_once '../models/productos.php';
+$productos = new Productos();
 switch ($option) {
 
-    case 'listarProductos':
-        if (isset($_GET['id_caja']) && isset($_GET['rolUsuario'])) {
-            $id_caja = intval($_GET['id_caja']);
-            $rolUsuario = intval($_GET['rolUsuario']);
-    
-            // Obtener productos
-            $productos = $compras->getProducts($id_caja, $rolUsuario);
-                $respuesta = [
-                'parametros_recibidos' => [
-                    'id_caja' => $id_caja,
-                    'rolUsuario' => $rolUsuario
-                ],
-                'productos' => $productos
-            ];
-            echo json_encode($respuesta);
-            
+    case 'listar':
+        // Verifica si el id_sede está en la sesión
+        if (isset($_SESSION['id_sede'])) {
+            $id_sede = $_SESSION['id_sede'];
+            $data = $productos->getProductsBySede($id_sede);  // Filtra los productos por sede
+            echo json_encode($data);
         } else {
-            echo json_encode(['error' => 'Faltan parámetros necesarios: id_caja o rolUsuario']);
+            echo json_encode(['tipo' => 'error', 'mensaje' => 'No se ha seleccionado ninguna sede.']);
         }
-        break;        
+        break;
     
-    case 'listarEmpresas':
-        $result = $compras->getEmpresas();
-        echo json_encode($result);
+
+    case 'save':
+        $barcode = $_POST['barcode'];
+        $nombre = $_POST['nombre'];
+        $precio_compra = $_POST['precio_compra'];
+        $precio_venta = $_POST['precio_venta'];
+        $stock = $_POST['stock'];
+        $imagen = $_FILES['imagen']['name'];
+        $id_proveedor = $_POST['id_proveedor'];
+        $id_product = $_POST['id_product'];
+        if ($id_product == '') {
+            $consult = $productos->comprobarBarcode($barcode);
+            if (empty($consult)) {
+                $result = $productos->saveProduct($barcode, $nombre, $precio_compra, $precio_venta, $stock, $imagen, $id_proveedor);
+                if ($result) {
+                    $res = array('tipo' => 'success', 'mensaje' => 'PRODUCTO REGISTRADO');
+                } else {
+                    $res = array('tipo' => 'error', 'mensaje' => 'ERROR AL AGREGAR');
+                }
+            } else {
+                $res = array('tipo' => 'error', 'mensaje' => 'EL BARCODE YA EXISTE');
+            }
+        } else {
+            $result = $productos->updateProduct($barcode, $nombre, $precio_compra, $precio_venta, $stock, $imagen, $id_proveedor, $id_product);
+            if ($result) {
+                $res = array('tipo' => 'success', 'mensaje' => 'PRODUCTO MODIFICADO');
+            } else {
+                $res = array('tipo' => 'error', 'mensaje' => 'ERROR AL MODIFICAR');
+            }
+        }
+        echo json_encode($res);
+        break;
+
+    case 'delete':
+        $id = $_GET['id'];
+        $data = $productos->deleteProducto($id);
+        if ($data) {
+            $res = array('tipo' => 'success', 'mensaje' => 'PRODUCTO ELIMINADO');
+        } else {
+            $res = array('tipo' => 'error', 'mensaje' => 'ERROR AL ELIMINAR');
+        }
+        echo json_encode($res);
         break;
         
-
-    case 'registrarCompra':
-        if (!isset($_SESSION['id_sede'])) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'No se ha seleccionado ninguna sede. Por favor abra una caja.']);
-            exit;
-        }
-    
-        $id_empresa = isset($_POST['id_empresa']) ? (int) $_POST['id_empresa'] : 0;
-        $sede = $_SESSION['id_sede'];
-        $precio_compra = isset($_POST['precio_compra']) ? (float) $_POST['precio_compra'] : 0.0;
-        $precio_venta = isset($_POST['precio_venta']) ? (float) $_POST['precio_venta'] : 0.0;
-        $cantidad = isset($_POST['cantidad']) ? (int) $_POST['cantidad'] : 0;
-        $estado = isset($_POST['estado']) ? (int) $_POST['estado'] : 1;
-        $descripcion = isset($_POST['descripcion']) ? $_POST['descripcion'] : '';
-        $barcode = isset($_POST['barcode']) ? $_POST['barcode'] : '';
-        $imagen = '';
-        $metodo_compra = isset($_POST['metodo_compra']) ? (int) $_POST['metodo_compra'] : 0;
-    
-        if (empty($id_empresa)) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Por favor seleccione una empresa.']);
-            break;
-        }
-    
-        if (isset($_FILES['imagen']) && $_FILES['imagen']['error'] == 0) {
-            if (!file_exists('../uploads')) {
-                mkdir('../uploads', 0777, true);
-            }
-            $imagen = 'uploads/' . basename($_FILES['imagen']['name']);
-            move_uploaded_file($_FILES['imagen']['tmp_name'], '../' . $imagen);
-        }
-    
-        $id_caja = isset($_SESSION['id_sede']) ? (int) $_SESSION['id_sede'] : 0;
-    
-        if ($id_caja === 0) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'No se ha podido obtener la sede del usuario.']);
-            break;
-        }
-    
-        $total = $precio_compra * $cantidad;
-    
-        $compraId = $compras->saveCompra($id_empresa, $total, null, $id_user, $estado, $id_caja, $metodo_compra);
-    
-        if (!$compraId) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Error al registrar la compra.']);
-            break;
-        }
-    
-        $productId = $compras->saveOrUpdateProduct($barcode, $descripcion, $id_empresa, $precio_compra, $precio_venta, $imagen, $cantidad, $estado, $id_caja);
-    
-        if (!$productId) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Error al guardar el producto.']);
-            break;
-        }
-    
-        $result = $compras->saveDetalle($productId, $compraId, $cantidad, $precio_compra);
-    
-        if (!$result) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Error al guardar el detalle de la compra.']);
-            break;
-        }
-    
-        echo json_encode(['tipo' => 'success', 'mensaje' => 'Compra registrada con éxito.']);
+    case 'edit':
+        $id = $_GET['id'];
+        $data = $productos->getProduct($id);
+        echo json_encode($data);
         break;
-    
-    case 'guardarDesdeModal':
-        $sede = isset($_POST['sede']) ? (int) $_POST['sede'] : 0;
-        $metodo = isset($_POST['metodo']) ? (int) $_POST['metodo'] : 0;
-        $estado = isset($_POST['estado']) ? (int) $_POST['estado'] : null; 
-        $id_user = $_SESSION['idusuario'];
-    
-        $id_empresa = isset($_POST['id_empresa']) ? (int) $_POST['id_empresa'] : 0;
-        $precio_compra = isset($_POST['precio_compra']) ? (float) $_POST['precio_compra'] : 0.0;
-        $precio_venta = isset($_POST['precio_venta']) ? (float) $_POST['precio_venta'] : 0.0;
-        $cantidad = isset($_POST['cantidad']) ? (int) $_POST['cantidad'] : 0;
-        $descripcion = isset($_POST['descripcion']) ? $_POST['descripcion'] : '';
-        $barcode = isset($_POST['barcode']) ? $_POST['barcode'] : '';
-        $imagen = isset($_FILES['imagen']) ? $_FILES['imagen'] : null;
-    
-        if ($sede === 0 || $metodo === 0) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Faltan datos del modal (sede o método) para guardar.']);
-            exit;
-        }
-    
-        $ruta_imagen = '';
-        if ($imagen && $imagen['error'] === 0) {
-            if (!file_exists('../uploads')) {
-                mkdir('../uploads', 0777, true);
-            }
-            $ruta_imagen = 'uploads/' . basename($imagen['name']);
-            move_uploaded_file($imagen['tmp_name'], '../' . $ruta_imagen);
-        }
-    
-        $compras->beginTransaction();
-    
-        try {
-            $total = $precio_compra * $cantidad;
-            $id_compra = $compras->saveCompra($id_empresa, $total, null, $id_user, $estado, $sede, $metodo);
-    
-            if (!$id_compra) {
-                throw new Exception('Error al guardar la compra.');
-            }
-    
-            $id_producto = $compras->saveOrUpdateProduct($barcode, $descripcion, $id_empresa, $precio_compra, $precio_venta, $ruta_imagen, $cantidad, $estado, $sede);
-    
-            if (!$id_producto) {
-                throw new Exception('Error al guardar el producto.');
-            }
-    
-            $detalle_guardado = $compras->saveDetalle($id_producto, $id_compra, $cantidad, $precio_compra);
-    
-            if (!$detalle_guardado) {
-                throw new Exception('Error al guardar el detalle de la compra.');
-            }
-    
-            $compras->commit();
-            echo json_encode(['tipo' => 'success', 'mensaje' => 'Datos del modal guardados con éxito.']);
-        } catch (Exception $e) {
-            $compras->rollBack();
-            echo json_encode(['tipo' => 'error', 'mensaje' => $e->getMessage()]);
-        }
-        break;
-             
-    
-    case 'cambiarEstado':
-        // Obtener los datos enviados por el frontend
-        $id_producto = isset($_POST['id_producto']) ? (int)$_POST['id_producto'] : 0;
-        $estado = isset($_POST['estado']) ? (int)$_POST['estado'] : null;
-        $barcode = isset($_POST['barcode']) ? trim($_POST['barcode']) : '';
-
-        // Validar los datos requeridos
-        if ($id_producto === 0 || is_null($estado) || empty($barcode)) {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Todos los campos son obligatorios para cambiar el estado.']);
-            exit;
-        }
-    
-        // Llamar al modelo para actualizar los datos
-        $result = $compras->updateEstadoProducto($id_producto, $estado, $barcode);
-    
-        // Verificar el resultado de la actualización
-        if ($result) {
-            echo json_encode(['tipo' => 'success', 'mensaje' => 'Estado del producto y la compra actualizado con éxito.']);
-        } else {
-            echo json_encode(['tipo' => 'error', 'mensaje' => 'Error al actualizar el estado del producto o la compra.']);
-        }
-        break;
-    
     default:
-    echo json_encode(['tipo' => 'error', 'mensaje' => 'Opción no válida.']);
-    error_log("Opción no válida: $option");
-    break;
+        # code...
+        break;
 }
+
